Redesign architecture
the new architecture concept is as follows:

train.py is the main script. Its mission is to read the configuration
file, set up the environment variables if needed, and to run() a
TrainManager. It should be used for the single-node, single-GPU case

launch.py is a top-level script that wraps torch.distributed.launch
to run train.py in a distributed setup.
https://github.com/pytorch/pytorch/blob/master/torch/distributed/launch.py

TrainManager is in charge of the generic experiment settings like
creating a timestamped folder, setting the loggers...
It uses a TrainerFactory to run() a Trainer defined in the
configuration file so it is not linked to any specific routine.

The Trainer is where all the machine learning thing happens. The
previous layers should have abstracted out all the distributed/logging/
path related stuff.

Trainers should take as arguments the TrainManager configuration, the
remaining argv not processed in the previous levels (this way we can
have Trainer-specific entries in the configuration file) and a
Tensorboard logger. And are expected to inherit the `TrainManager`
namespace for logging.

Trainers can use the ModelFactory to run under different architectures
without modifying its code.