Add ONNX  exporter
On testing, it was found that the ONNX to TensorFlow conversion process
is not able to handle multiple outputs.
Thus, the models are now expected to return a single output (logits) for
training and validation.
An auxiliary Sigmoid layer is attached to the model to compute the
activations and also to export the final deployment ONNX model.